Implementing feedback from unknown hero (thank you)
Changes how RV events are bound to the ReviewMenu class.

Instead of binding events within the `setup_listeners` method,
the events are now bound during the class's initialization.

This approach uses bindings defined within the `init` method to
manage RV event subscriptions, and the previous `setup_listeners`
method is commented out.

The change aims to ensure that event bindings are managed in a
centralized location.